+++ src/lib/core/input.rs
@@ -6,19 +6,19 @@ use core::State;
 /// An input that is is guarenteed to be between two values
 pub trait Bounded: Tree + Send + Sync {
     /// Get the input value between the bounds given by `get_bounds`
-    fn get(&mut self, state: &State) -> f32;
+    fn get(&mut self, state: &State) -> f64;
 
     /// Get the lower and upper bounds (respectively) of the number returned by
     /// `get`
-    fn get_bounds(&self) -> (f32, f32);
+    fn get_bounds(&self) -> (f64, f64);
 
     /// Get the value bounded between two values
     fn get_with_bounds(
         &mut self,
         state: &State,
-        lower: f32,
-        upper: f32,
-    ) -> f32
+        lower: f64,
+        upper: f64,
+    ) -> f64
     {
         let value = self.get(state);
         let (self_lower, self_upper) = self.get_bounds();
